Mosip-44078: Sonar duplication
