+++ pmp-ui-v2/src/pages/partner/authenticationServices/EditOidcClient.js
@@ -13,7 +13,9 @@ import {
     createDropdownData,
     createRequest,
     isOidcClientAdditionalInfoRequired,
-    buildClientNameLangMap
+    buildClientNameLangMap,
+    createOidcClientEntry, findAvailableOidcLanguage, validateOidcEntryText,
+    getOidcPlaceholderForLanguage, getAvailableOidcLanguages, getAvailableLanguagesForClientNameLangMap
 } from "../../../utils/AppUtils";
 import LoadingIcon from "../../common/LoadingIcon";
 import ErrorMessage from "../../common/ErrorMessage";
@@ -26,6 +28,7 @@ import somethingWentWrongIcon from '../../../svg/something_went_wrong_icon.svg';
 import expandToggleIcon from '../../../svg/expand_toggle_icon.svg';
 import TextInputComponentWithDeleteButton from '../../common/fields/TextInputComponentWithDeleteButton';
 import { getAppConfig } from '../../../services/ConfigService';
+import OidcClientAdditionalInfoSection from '../../common/OidcClientAdditionalInfoSection';
 
 function EditOidcClient() {
     const { t } = useTranslation();
@@ -339,37 +342,6 @@ function EditOidcClient() {
         setErrorMsg("");
     };
 
-    // Helper function to create a new entry
-    const createNewEntry = (language, type) => {
-        const baseId = crypto.randomUUID();
-
-        let uniqueId;
-        
-        switch (type) {
-            case 'clientName':
-                uniqueId = `oidc_name_${baseId}`;
-                break;
-            case 'purposeTitle':
-                uniqueId = `purpose_title_${baseId}`;
-                break;
-            case 'purposeSubtitle':
-                uniqueId = `purpose_subtitle_${baseId}`;
-                break;
-            default:
-                uniqueId = `entry_${type}_${baseId}`;
-        }
-        
-        return {
-            id: uniqueId,
-            language: language,
-            text: ''
-        };
-    };
-
-    // Helper function to find available language
-    const findAvailableLanguage = (usedLanguages) => {
-        return languageDropdownData.find(lang => !usedLanguages.includes(lang.fieldValue));
-    };
 
     // Client name handler
     const handleClientNameChange = (value) => {
@@ -395,27 +367,10 @@ function EditOidcClient() {
             lang.fieldValue !== 'default' && !usedLanguages.includes(lang.fieldValue)
         );
         const availableLang = availableLangs[0];
-        const newEntry = createNewEntry(availableLang?.fieldValue, 'clientName');
+        const newEntry = createOidcClientEntry(availableLang?.fieldValue, 'clientName');
         setClientNameLangMapEntries([...clientNameLangMapEntries, newEntry]);
     };
 
-    const validateEntryText = (value, entry, requiredErrorKey, errors, setErrors) => {
-        const newErrors = { ...errors };
-        let inputError = "";
-
-        validateInputRegex(value, (error) => {
-            inputError = error;
-        }, t);
-
-        if (entry.text.trim() === '' && entry.text !== '') {
-            newErrors[entry.id] = t(requiredErrorKey);
-        } else if (inputError) {
-            newErrors[entry.id] = inputError;
-        } else {
-            delete newErrors[entry.id];
-        }
-        setErrors(newErrors);
-    };
 
     const updateClientNameLangMapEntry = (id, field, value) => {
         const updated = clientNameLangMapEntries.map(entry =>
@@ -425,7 +380,7 @@ function EditOidcClient() {
 
         if (field === 'text') {
             const entry = updated.find(e => e.id === id);
-            validateEntryText(value, entry, 'createOidcClient.clientNameRequired', clientNameLangMapErrors, setClientNameLangMapErrors);
+            validateOidcEntryText(value, entry, 'createOidcClient.clientNameRequired', clientNameLangMapErrors, setClientNameLangMapErrors, t);
         }
     };
 
@@ -437,38 +392,6 @@ function EditOidcClient() {
         setClientNameLangMapErrors(errors);
     };
 
-    const getAvailableLanguagesForClientNameLangMap = (currentEntryId) => {
-        const currentEntry = clientNameLangMapEntries.find(e => e.id === currentEntryId);
-        const currentLanguage = currentEntry ? currentEntry.language : '';
-        const usedLanguages = clientNameLangMapEntries
-            .filter(e => e.id !== currentEntryId && e.language)
-            .map(e => e.language);
-        return languageDropdownData.filter(lang =>
-            lang.fieldValue !== 'default' && (!usedLanguages.includes(lang.fieldValue) || lang.fieldValue === currentLanguage)
-        );
-    };
-
-    // Get placeholder based on language code and field type
-    const getPlaceholderForLanguage = (languageCode, fieldType) => {
-        if (!languageCode || languageCode === 'default') {
-            const fallbackKey = `createOidcClient.enter${fieldType}Default`;
-            return t(fallbackKey);
-        }
-        
-        const langCode = languageCode.toLowerCase();
-        
-        // Use the new translation keys that have text in the target language
-        const placeholderKey = `createOidcClient.enter${fieldType}In${langCode.charAt(0).toUpperCase() + langCode.slice(1)}`;
-        const fallbackKey = `createOidcClient.enter${fieldType}Default`;
-        
-        // Get translation (all translation files have the same keys with target language text)
-        let placeholder = t(placeholderKey);
-        if (placeholder === placeholderKey) {
-            placeholder = t(fallbackKey);
-        }
-        
-        return placeholder;
-    };
 
     const handleLogoUrlChange = (value) => {
         setInvalidLogoUrl(validateUrl(null, value, 2048, [], t));
@@ -569,8 +492,8 @@ function EditOidcClient() {
         const usedLanguages = purposeTitleEntries.map(e => e.language).filter(lang => lang);
         const availableLang = purposeTitleEntries.length === 0
             ? languageDropdownData.find(lang => lang.fieldValue === 'default')
-            : findAvailableLanguage(usedLanguages);
-        const newEntry = createNewEntry(availableLang?.fieldValue, 'purposeTitle');
+            : findAvailableOidcLanguage(usedLanguages, languageDropdownData);
+        const newEntry = createOidcClientEntry(availableLang?.fieldValue, 'purposeTitle');
         const updated = [...purposeTitleEntries, newEntry];
         setPurposeTitleEntries(updated);
         validatePurposeDefaultRequirement(updated, 'title');
@@ -580,8 +503,8 @@ function EditOidcClient() {
         const usedLanguages = purposeSubtitleEntries.map(e => e.language).filter(lang => lang);
         const availableLang = purposeSubtitleEntries.length === 0
             ? languageDropdownData.find(lang => lang.fieldValue === 'default')
-            : findAvailableLanguage(usedLanguages);
-        const newEntry = createNewEntry(availableLang?.fieldValue, 'purposeSubtitle');
+            : findAvailableOidcLanguage(usedLanguages, languageDropdownData);
+        const newEntry = createOidcClientEntry(availableLang?.fieldValue, 'purposeSubtitle');
         const updated = [...purposeSubtitleEntries, newEntry];
         setPurposeSubtitleEntries(updated);
         validatePurposeDefaultRequirement(updated, 'subtitle');
@@ -595,7 +518,7 @@ function EditOidcClient() {
 
         if (field === 'text') {
             const entry = updated.find(e => e.id === id);
-            validateEntryText(value, entry, 'createOidcClient.purposeTitleRequired', purposeTitleErrors, setPurposeTitleErrors);
+            validateOidcEntryText(value, entry, 'createOidcClient.purposeTitleRequired', purposeTitleErrors, setPurposeTitleErrors, t);
         }
         validatePurposeDefaultRequirement(updated, 'title');
     };
@@ -608,7 +531,7 @@ function EditOidcClient() {
 
         if (field === 'text') {
             const entry = updated.find(e => e.id === id);
-            validateEntryText(value, entry, 'createOidcClient.purposeSubtitleRequired', purposeSubtitleErrors, setPurposeSubtitleErrors);
+            validateOidcEntryText(value, entry, 'createOidcClient.purposeSubtitleRequired', purposeSubtitleErrors, setPurposeSubtitleErrors, t);
         }
         validatePurposeDefaultRequirement(updated, 'subtitle');
     };
@@ -641,18 +564,6 @@ function EditOidcClient() {
         delete errors[id];
         setPurposeSubtitleErrors(errors);
         validatePurposeDefaultRequirement(updated, 'subtitle');
-    };
-
-    const getAvailableLanguages = (currentEntryId, type) => {
-        const entries = type === 'title' ? purposeTitleEntries : purposeSubtitleEntries;
-        const currentEntry = entries.find(e => e.id === currentEntryId);
-        const currentLanguage = currentEntry ? currentEntry.language : '';
-        const usedLanguages = entries
-            .filter(e => e.id !== currentEntryId && e.language)
-            .map(e => e.language);
-        return languageDropdownData.filter(lang =>
-            !usedLanguages.includes(lang.fieldValue) || lang.fieldValue === currentLanguage
-        );
     };
 
     // Build purpose title/subtitle lang map with @none mapping
@@ -1116,7 +1027,7 @@ function EditOidcClient() {
                                                                     ) : (
                                                                         <div className="bg-white border border-neutral-300 shadow-sm rounded-md p-4">
                                                                             {clientNameLangMapEntries.map((entry, index) => {
-                                                                                const availableLangs = getAvailableLanguagesForClientNameLangMap(entry.id);
+                                                                                const availableLangs = getAvailableLanguagesForClientNameLangMap(entry.id, clientNameLangMapEntries, languageDropdownData);
                                                                                 return (
                                                                                     <div key={index} className="flex mb-2">
                                                                                         <div className="w-1/3">
@@ -1135,7 +1046,7 @@ function EditOidcClient() {
                                                                                                 value={entry.text}
                                                                                                 onChange={(e) => updateClientNameLangMapEntry(entry.id, 'text', e.target.value)}
                                                                                                 onDelete={() => deleteClientNameLangMapEntry(entry.id)}
-                                                                                                placeholder={getPlaceholderForLanguage(entry.language, 'NameForOidcClient')}
+                                                                                                placeholder={getOidcPlaceholderForLanguage(entry.language, 'NameForOidcClient', t)}
                                                                                                 id={`client_name_lang_map_text_${index + 1}`}
                                                                                                 maxLength={256}
                                                                                                 showDelete={clientNameLangMapEntries.length > 0}
@@ -1242,301 +1153,39 @@ function EditOidcClient() {
                                                 </div>
 
                                                 {/* Additional Information Section */}
-                                                <div className="bg-snow-white px-7 py-4 mt-[1.5%] mb-4 rounded-lg shadow-md">
-                                                    <div
-                                                        className="flex items-center justify-between cursor-pointer"
-                                                        onClick={() => setIsAdditionalInfoExpanded(!isAdditionalInfoExpanded)}
-                                                        role="button"
-                                                        tabIndex="0"
-                                                        onKeyDown={(e) => onPressEnterKey(e, () => setIsAdditionalInfoExpanded(!isAdditionalInfoExpanded))}
-                                                    >
-                                                        <h3 className="text-lg font-semibold text-dark-blue">{t('createOidcClient.additionalInformation')}</h3>
-                                                        <img src={expandToggleIcon} alt="Toggle" className={`w-7 h-7 transform transition-transform ${isAdditionalInfoExpanded ? 'rotate-180' : ''}`} />
-                                                    </div>
-
-                                                    {isAdditionalInfoExpanded && (
-                                                        <div className="flex flex-col space-y-4 mt-3">
-                                                            <div className="border-b border-gray-200 mb-3"></div>
-                                                            
-                                                            {/* Toggles - Side by Side */}
-                                                            <div className="flex justify-between space-x-4">
-                                                                {/* Forgot Password Banner Toggle */}
-                                                                <div className="flex flex-col w-[48%]">
-                                                                    <div className={`flex items-center ${isLoginLanguageRTL ? 'flex-row-reverse justify-end' : 'justify-start'} gap-3`}>
-                                                                        <div className="flex items-center">
-                                                                            <label htmlFor="forgot_password_banner_toggle" className={`text-dark-blue text-sm font-semibold ${isLoginLanguageRTL ? "mr-1" : "ml-1"}`}>
-                                                                                {t('createOidcClient.forgotPasswordBanner')}
-                                                                            </label>
-                                                                            <Information infoKey={t('createOidcClient.forgotPasswordBannerTooltip')} id='forgot_password_banner_info' />
-                                                                        </div>
-                                                                        <button
-                                                                            type="button"
-                                                                            role="switch"
-                                                                            aria-checked={forgotPasswordBanner}
-                                                                            aria-label={t('createOidcClient.forgotPasswordBanner')}
-                                                                            className="relative inline-flex items-center cursor-pointer flex-shrink-0 focus:outline focus:outline-2 focus:outline-[#1447B2] focus:outline-offset-2 rounded"
-                                                                            onClick={() => additionalConfigRequired && setForgotPasswordBanner(!forgotPasswordBanner)}
-                                                                            disabled={!additionalConfigRequired}
-                                                                        >
-                                                                            <input
-                                                                                type="checkbox"
-                                                                                checked={forgotPasswordBanner}
-                                                                                onChange={(e) => additionalConfigRequired && setForgotPasswordBanner(e.target.checked)}
-                                                                                className="sr-only peer focus:outline-none"
-                                                                                id="forgot_password_banner_toggle"
-                                                                                tabIndex={-1}
-                                                                            />
-                                                                            <div className={`relative w-9 h-5 rounded-full transition-colors duration-200 ease-in-out ${forgotPasswordBanner ? 'bg-[#1447B2]' : 'bg-neutral-100'}`}>
-                                                                                <div className={`absolute top-0.5 left-0.5 w-4 h-4 bg-white rounded-full shadow-sm transform transition-transform duration-200 ease-in-out ${forgotPasswordBanner ? 'translate-x-4' : ''}`}></div>
-                                                                            </div>
-                                                                        </button>
-                                                                    </div>
-                                                                </div>
-
-                                                                {/* SignUp Banner Toggle */}
-                                                                <div className="flex flex-col w-[48%]">
-                                                                    <div className={`flex items-center ${isLoginLanguageRTL ? 'flex-row-reverse justify-end' : 'justify-start'} gap-3`}>
-                                                                        <div className="flex items-center">
-                                                                            <label htmlFor="signup_banner_toggle" className={`text-dark-blue text-sm font-semibold ${isLoginLanguageRTL ? "mr-1" : "ml-1"}`}>
-                                                                                {t('createOidcClient.signUpBanner')}
-                                                                            </label>
-                                                                            <Information infoKey={t('createOidcClient.signUpBannerTooltip')} id='signup_banner_info' />
-                                                                        </div>
-                                                                        <button
-                                                                            type="button"
-                                                                            role="switch"
-                                                                            aria-checked={signUpBanner}
-                                                                            aria-label={t('createOidcClient.signUpBanner')}
-                                                                            className="relative inline-flex items-center cursor-pointer flex-shrink-0 focus:outline focus:outline-2 focus:outline-[#1447B2] focus:outline-offset-2 rounded"
-                                                                            onClick={() => additionalConfigRequired && setSignUpBanner(!signUpBanner)}
-                                                                            disabled={!additionalConfigRequired}
-                                                                        >
-                                                                            <input
-                                                                                type="checkbox"
-                                                                                checked={signUpBanner}
-                                                                                onChange={(e) => additionalConfigRequired && setSignUpBanner(e.target.checked)}
-                                                                                className="sr-only peer focus:outline-none"
-                                                                                id="signup_banner_toggle"
-                                                                                tabIndex={-1}
-                                                                            />
-                                                                            <div className={`relative w-9 h-5 rounded-full transition-colors duration-200 ease-in-out ${signUpBanner ? 'bg-[#1447B2]' : 'bg-neutral-100'}`}>
-                                                                                <div className={`absolute top-0.5 left-0.5 w-4 h-4 bg-white rounded-full shadow-sm transform transition-transform duration-200 ease-in-out ${signUpBanner ? 'translate-x-4' : ''}`}></div>
-                                                                            </div>
-                                                                        </button>
-                                                                    </div>
-                                                                </div>
-                                                            </div>
-
-                                                            {/* Consent Expiry Duration and User Info Response Type - Side by Side */}
-                                                            <div className="flex flex-row justify-between space-x-4">
-                                                                {/* Consent Expiry Duration */}
-                                                                <div className="flex flex-col w-[48%]">
-                                                                    <label id="consent_expiry_label" className={`flex items-center text-dark-blue text-sm font-semibold mb-1 ${isLoginLanguageRTL ? "mr-1" : "ml-1"}`}>
-                                                                        {t('createOidcClient.consentExpiryDuration')}<span className="text-crimson-red mx-1">*</span>
-                                                                        <Information infoKey={t('createOidcClient.consentExpiryDurationTooltip')} id='consent_expiry_info' />
-                                                                    </label>
-                                                                    <input
-                                                                        id="consent_expiry_input"
-                                                                        value={consentExpiry}
-                                                                        onChange={(e) => handleConsentExpiryChange(e.target.value)}
-                                                                        className={`h-10 px-2 py-3 border border-[#707070] rounded-md text-base text-dark-blue ${additionalConfigRequired ? "bg-white" : "bg-platinum-gray"} leading-tight focus:outline-none focus:shadow-outline w-full`}
-                                                                        placeholder={additionalConfigRequired && t('createOidcClient.consentExpiryPlaceholder')}
-                                                                        readOnly={!additionalConfigRequired}
-                                                                    />
-                                                                    {consentExpiryError && <span id="consent_expiry_error" className="text-sm text-crimson-red font-semibold mt-1">{consentExpiryError}</span>}
-                                                                </div>
-
-                                                                {/* User Info Response Type */}
-                                                                <div className="flex flex-col w-[48%]">
-                                                                    <DropdownComponent
-                                                                        fieldName='userInfoResponseType'
-                                                                        dropdownDataList={userInfoResponseTypeDropdownData}
-                                                                        onDropDownChangeEvent={handleUserInfoResponseTypeChange}
-                                                                        fieldNameKey='createOidcClient.userInfoResponseType'
-                                                                        placeHolderKey='createOidcClient.selectUserInfoResponseType'
-                                                                        selectedDropdownValue={userInfoResponseType}
-                                                                        styleSet={styles}
-                                                                        addInfoIcon={true}
-                                                                        infoKey={t('createOidcClient.userInfoResponseTypeTooltip')}
-                                                                        isPlaceHolderPresent={true}
-                                                                        id='user_info_response_type'
-                                                                        disabled={!additionalConfigRequired} />
-                                                                </div>
-                                                            </div>
-
-                                                            {/* Purpose Type - Full Width */}
-                                                            <div className="flex flex-col w-[48%] mb-2">
-                                                                <DropdownComponent
-                                                                    fieldName='purposeType'
-                                                                    dropdownDataList={purposeTypeDropdownData}
-                                                                    onDropDownChangeEvent={handlePurposeTypeChange}
-                                                                    fieldNameKey='createOidcClient.purposeType'
-                                                                    placeHolderKey='createOidcClient.selectPurposeType'
-                                                                    selectedDropdownValue={purposeType}
-                                                                    styleSet={styles}
-                                                                    addInfoIcon={true}
-                                                                    infoKey={t('createOidcClient.purposeTypeTooltip')}
-                                                                    isPlaceHolderPresent={true}
-                                                                    id='purpose_type'
-                                                                    disabled={!additionalConfigRequired} />
-                                                            </div>
-
-                                                            {/* Purpose Title - shown only when Purpose Type is selected */}
-                                                            {purposeType && (
-                                                                <div className="flex flex-col">
-                                                                    <label id="purpose_title_label" className={`flex items-center text-dark-blue text-sm font-semibold mb-2 ${isLoginLanguageRTL ? "mr-1" : "ml-1"}`}>
-                                                                        {t('createOidcClient.purposeTitle')}
-                                                                        <Information infoKey={t('createOidcClient.purposeTitleTooltip')} id='purpose_title_info' />
-                                                                    </label>
-                                                                    {purposeTitleEntries.length === 0 ? (
-                                                                        <div className="bg-white border border-neutral-200 rounded-md p-8 flex flex-col items-center justify-center min-h-[120px]">
-                                                                            <button
-                                                                                type="button"
-                                                                                id="add_purpose_title_language"
-                                                                                className="bg-[#1447b2] text-white font-semibold text-sm px-6 py-2 rounded-md cursor-pointer hover:bg-[#0f3a8a] transition-colors"
-                                                                                tabIndex="0"
-                                                                                onKeyDown={(e) => onPressEnterKey(e, addPurposeTitleEntry)}
-                                                                                onClick={addPurposeTitleEntry}
-                                                                            >
-                                                                                {t('createOidcClient.addTitle')}
-                                                                            </button>
-                                                                            <p className="text-gray-400 text-sm mt-2 text-center">
-                                                                                {t('createOidcClient.addTitleHelperText')}
-                                                                            </p>
-                                                                        </div>
-                                                                    ) : (
-                                                                        <div className="bg-white border border-neutral-300 shadow-sm rounded-md p-4">
-                                                                            {purposeTitleEntries.map((entry, index) => {
-                                                                                const availableLangs = getAvailableLanguages(entry.id, 'title');
-                                                                                return (
-                                                                                    <div key={index} className="flex mb-2">
-                                                                                        <div className="w-1/3">
-                                                                                            <DropdownComponent
-                                                                                                fieldName={`purposeTitleLang_${index + 1}`}
-                                                                                                dropdownDataList={availableLangs}
-                                                                                                onDropDownChangeEvent={(field, value) => updatePurposeTitleEntry(entry.id, 'language', value)}
-                                                                                                fieldNameKey=""
-                                                                                                placeHolderKey="createOidcClient.selectLanguage"
-                                                                                                selectedDropdownValue={entry.language}
-                                                                                                styleSet={styles}
-                                                                                                id={`purpose_title_lang_${index + 1}`} />
-                                                                                        </div>
-                                                                                        <div className={`w-full mt-1 ${isLoginLanguageRTL ? 'mr-5' : 'ml-5'}`}>
-                                                                                            <TextInputComponentWithDeleteButton
-                                                                                                value={entry.text}
-                                                                                                onChange={(e) => updatePurposeTitleEntry(entry.id, 'text', e.target.value)}
-                                                                                                onDelete={() => deletePurposeTitleEntry(entry.id)}
-                                                                                                placeholder={getPlaceholderForLanguage(entry.language, 'PurposeTitle')}
-                                                                                                id={`purpose_title_text_${index + 1}`}
-                                                                                                showDelete={purposeTitleEntries.length > 0}
-                                                                                                errorMessage={purposeTitleErrors[entry.id]}
-                                                                                                isRTL={isLangRTL(entry.language)}
-                                                                                                languageCode={entry.language}
-                                                                                            />
-                                                                                        </div>
-                                                                                    </div>
-                                                                                );
-                                                                            })}
-                                                                            {purposeTitleEntries.length < languageDropdownData.length && (
-                                                                                <div
-                                                                                    role="button"
-                                                                                    id="add_purpose_title_language"
-                                                                                    className="text-[#1447b2] font-bold text-xs w-fit cursor-pointer"
-                                                                                    tabIndex="0"
-                                                                                    onKeyDown={(e) => onPressEnterKey(e, addPurposeTitleEntry)}
-                                                                                    onClick={addPurposeTitleEntry}
-                                                                                >
-                                                                                    <span className="text-lg text-center">+</span>
-                                                                                    <span>{t('createOidcClient.addNew')}</span>
-                                                                                </div>
-                                                                            )}
-                                                                        </div>
-                                                                    )}
-                                                                    {purposeTitleDefaultError && (
-                                                                        <span className="text-sm text-crimson-red font-semibold mt-2">{purposeTitleDefaultError}</span>
-                                                                    )}
-                                                                </div>
-                                                            )}
-
-                                                            {/* Purpose Subtitle - shown only when Purpose Type is selected */}
-                                                            {purposeType && (
-                                                                <div className="flex flex-col">
-                                                                    <label id="purpose_subtitle_label" className={`flex items-center text-dark-blue text-sm font-semibold mb-2 ${isLoginLanguageRTL ? "mr-1" : "ml-1"}`}>
-                                                                        {t('createOidcClient.purposeSubtitle')}
-                                                                        <Information infoKey={t('createOidcClient.purposeSubtitleTooltip')} id='purpose_subtitle_info' />
-                                                                    </label>
-                                                                    {purposeSubtitleEntries.length === 0 ? (
-                                                                        <div className="bg-white border border-neutral-200 rounded-md p-8 flex flex-col items-center justify-center min-h-[120px]">
-                                                                            <button
-                                                                                type="button"
-                                                                                id="add_purpose_subtitle_language"
-                                                                                className="bg-[#1447b2] text-white font-semibold text-sm px-6 py-2 rounded-md cursor-pointer hover:bg-[#0f3a8a] transition-colors"
-                                                                                tabIndex="0"
-                                                                                onKeyDown={(e) => onPressEnterKey(e, addPurposeSubtitleEntry)}
-                                                                                onClick={addPurposeSubtitleEntry}
-                                                                            >
-                                                                                {t('createOidcClient.addSubtitle')}
-                                                                            </button>
-                                                                            <p className="text-gray-400 text-sm mt-2 text-center">
-                                                                                {t('createOidcClient.addSubtitleHelperText')}
-                                                                            </p>
-                                                                        </div>
-                                                                    ) : (
-                                                                        <div className="bg-white border border-neutral-300 shadow-sm rounded-md p-4">
-                                                                            {purposeSubtitleEntries.map((entry, index) => {
-                                                                                const availableLangs = getAvailableLanguages(entry.id, 'subtitle');
-                                                                                return (
-                                                                                    <div key={index} className="flex mb-2">
-                                                                                        <div className="w-1/3">
-                                                                                            <DropdownComponent
-                                                                                                fieldName={`purposeSubtitleLang_${index + 1}`}
-                                                                                                dropdownDataList={availableLangs}
-                                                                                                onDropDownChangeEvent={(field, value) => updatePurposeSubtitleEntry(entry.id, 'language', value)}
-                                                                                                fieldNameKey=""
-                                                                                                placeHolderKey="createOidcClient.selectLanguage"
-                                                                                                selectedDropdownValue={entry.language}
-                                                                                                styleSet={styles}
-                                                                                                id={`purpose_subtitle_lang_${index + 1}`} />
-                                                                                        </div>
-                                                                                        <div className={`w-full mt-1 ${isLoginLanguageRTL ? 'mr-5' : 'ml-5'}`}>
-                                                                                            <TextInputComponentWithDeleteButton
-                                                                                                value={entry.text}
-                                                                                                onChange={(e) => updatePurposeSubtitleEntry(entry.id, 'text', e.target.value)}
-                                                                                                onDelete={() => deletePurposeSubtitleEntry(entry.id)}
-                                                                                                placeholder={getPlaceholderForLanguage(entry.language, 'PurposeSubtitle')}
-                                                                                                id={`purpose_subtitle_text_${index + 1}`}
-                                                                                                showDelete={purposeSubtitleEntries.length > 0}
-                                                                                                errorMessage={purposeSubtitleErrors[entry.id]}
-                                                                                                isRTL={isLangRTL(entry.language)}
-                                                                                                languageCode={entry.language}
-                                                                                            />
-                                                                                        </div>
-                                                                                    </div>
-                                                                                );
-                                                                            })}
-                                                                            {purposeSubtitleEntries.length < languageDropdownData.length && (
-                                                                                <div
-                                                                                    role="button"
-                                                                                    id="add_purpose_subtitle_language"
-                                                                                    className="text-[#1447b2] font-bold text-xs w-fit cursor-pointer"
-                                                                                    tabIndex="0"
-                                                                                    onKeyDown={(e) => onPressEnterKey(e, addPurposeSubtitleEntry)}
-                                                                                    onClick={addPurposeSubtitleEntry}
-                                                                                >
-                                                                                    <span className="text-lg text-center">+</span>
-                                                                                    <span>{t('createOidcClient.addNew')}</span>
-                                                                                </div>
-                                                                            )}
-                                                                        </div>
-                                                                    )}
-                                                                    {purposeSubtitleDefaultError && (
-                                                                        <span className="text-sm text-crimson-red font-semibold mt-2">{purposeSubtitleDefaultError}</span>
-                                                                    )}
-                                                                </div>
-                                                            )}
-                                                        </div>
-                                                    )}
-                                                </div>
+                                                <OidcClientAdditionalInfoSection
+                                                  isAdditionalInfoExpanded={isAdditionalInfoExpanded}
+                                                  setIsAdditionalInfoExpanded={setIsAdditionalInfoExpanded}
+                                                  isLoginLanguageRTL={isLoginLanguageRTL}
+                                                  additionalConfigRequired={additionalConfigRequired}
+                                                  forgotPasswordBanner={forgotPasswordBanner}
+                                                  setForgotPasswordBanner={setForgotPasswordBanner}
+                                                  signUpBanner={signUpBanner}
+                                                  setSignUpBanner={setSignUpBanner}
+                                                  consentExpiry={consentExpiry}
+                                                  handleConsentExpiryChange={handleConsentExpiryChange}
+                                                  consentExpiryError={consentExpiryError}
+                                                  userInfoResponseType={userInfoResponseType}
+                                                  handleUserInfoResponseTypeChange={handleUserInfoResponseTypeChange}
+                                                  userInfoResponseTypeDropdownData={userInfoResponseTypeDropdownData}
+                                                  purposeType={purposeType}
+                                                  handlePurposeTypeChange={handlePurposeTypeChange}
+                                                  purposeTypeDropdownData={purposeTypeDropdownData}
+                                                  purposeTitleEntries={purposeTitleEntries}
+                                                  addPurposeTitleEntry={addPurposeTitleEntry}
+                                                  updatePurposeTitleEntry={updatePurposeTitleEntry}
+                                                  deletePurposeTitleEntry={deletePurposeTitleEntry}
+                                                  purposeTitleErrors={purposeTitleErrors}
+                                                  purposeTitleDefaultError={purposeTitleDefaultError}
+                                                  purposeSubtitleEntries={purposeSubtitleEntries}
+                                                  addPurposeSubtitleEntry={addPurposeSubtitleEntry}
+                                                  updatePurposeSubtitleEntry={updatePurposeSubtitleEntry}
+                                                  deletePurposeSubtitleEntry={deletePurposeSubtitleEntry}
+                                                  purposeSubtitleErrors={purposeSubtitleErrors}
+                                                  purposeSubtitleDefaultError={purposeSubtitleDefaultError}
+                                                  languageDropdownData={languageDropdownData}
+                                                  styles={styles}
+                                                />
                                             </form>
                                         </div>
                                         <div className="pb-3 pt-6 px-4 bg-snow-white mt-[1.5%] rounded-lg shadow-md">
